backend/pyproject.toml
[project]
name = "edu-fi"
version = "0.1.0"
description = "Education platform backend API"
authors = [
    { name = "Edu-Fi Team" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "alembic (>=1.15.1,<2.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
pytest-cov = "^6.0.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"

